+++ src/_core/mifrost/core/encoders/hetero/semantic_hgraph_encoder.hpp
@@ -104,6 +104,7 @@ class MIFROST_API SemanticHGraphEncoderEngine {
    [[nodiscard]] const std::vector< SemanticPredicateSpec >& get_predicates() const;
    [[nodiscard]] const std::vector< SemanticActionSpec >& get_actions() const;
    [[nodiscard]] const std::map< std::string, int >& get_relation_arities() const;
+   void update_relations(std::map< std::string, int > relation_arities);
 
   private:
    friend class SemanticSuccessorHGraphEncoderEngine;